Stop normalizing refs to cid_v1 - store them as-referenced
Normalization seemed like a straightforwad thing to do at the time: it allowed
to instantly know which logical dags are part of which other largr dags, and do
the right thing when aggregating. Everything even worked for nearly two months
with no failure.

Then someone uploaded a whole bunch of identical blocks, referenced by v0 and v1
in different dags: 3.5GiB of such blocks, to be precise. When these dags ended
up aggregated in a superdag, things started failing during .car generation, as
the .car spec explicitly does not de-duplicate blocks based on multihash. The
aggregator would correctly observe that some dags are part of larger ones,
exclude them from the initial selection/size estimation, then when the aggregate
candidate groups are put together, turn around and re-inject all "already included
parts" assuming that the cost will be a few megabytes of overhead at most. That
assumption no longer held :cryingbear:

Thus change the tracking of refs to use its native encoding ( be it cidv1-b32 or
cidv0 ), while keeping everything else cidv1-normalized as before. This breaks
the ability to tell if a particular v0-referenced-dag is a part of a larger one.
As a result dag-size estimation and thus aggregate-ceiling-bytes will again work
as expected.